DOC: Rewrite nbgen
Using nbformat and nbconvert instead of our own custom
tools.

COMPAT: use io.open and UTF-8 encoding
